+++ src/Examples/TypedEncoding.hs
@@ -4,7 +4,9 @@ module Examples.TypedEncoding (
    -- * Conversions between encodings 
    , module Examples.TypedEncoding.Conversions
    -- * Adding new encodings, error handling      
-   , module Examples.TypedEncoding.DiySignEncoding
+   , module Examples.TypedEncoding.Instances.DiySignEncoding
+   -- * Examples of "do-" encodings that perform some string transformation such as title-casing     
+   , module Examples.TypedEncoding.Instances.Do.Sample
    -- * Converting other types to and from encoded strings 
    , module Examples.TypedEncoding.ToEncString
    -- * Modifying encoded payload    
@@ -13,6 +15,7 @@ module Examples.TypedEncoding (
 
 import           Examples.TypedEncoding.Overview
 import           Examples.TypedEncoding.Conversions     
-import           Examples.TypedEncoding.DiySignEncoding  
+import           Examples.TypedEncoding.Instances.DiySignEncoding  
+import           Examples.TypedEncoding.Instances.Do.Sample
 import           Examples.TypedEncoding.Unsafe  
 import           Examples.TypedEncoding.ToEncString