+++ src/Examples/TypedEncoding/Overview.hs
@@ -16,7 +16,7 @@
 --
 -- * "Data.TypedEncoding.Instances.Enc.Base64"
 -- * "Data.TypedEncoding.Instances.Restriction.ASCII"
--- * "Data.TypedEncoding.Instances.Do.Sample"
+-- * "Examples.TypedEncoding.Instances.Do.Sample"
 --
 
 module Examples.TypedEncoding.Overview where
@@ -24,7 +24,7 @@ module Examples.TypedEncoding.Overview where
 import           Data.TypedEncoding
 import           Data.TypedEncoding.Instances.Enc.Base64 ()
 import           Data.TypedEncoding.Instances.Restriction.ASCII ()
-import           Data.TypedEncoding.Instances.Do.Sample
+import           Examples.TypedEncoding.Instances.Do.Sample
  
 import qualified Data.ByteString as B
 import qualified Data.Text as T
@@ -151,7 +151,7 @@ helloB64B64RecoveredErr = recreateFAll . toEncoding () $ "SGVsbG8gV29ybGQ="
 -- * "do-" Encodings
 
 -- |
--- "do-UPPER" (from 'Data.TypedEncoding.Instances.Do.Sample' module) encoding applied to "Hello World"
+-- "do-UPPER" (from 'Examples.TypedEncoding.Instances.Do.Sample' module) encoding applied to "Hello World"
 --
 -- Notice a namespace thing going on, "enc-" is encoding, "do-" is some transformation. 
 -- These are typically not reversible, some could be recoverable.